client-services/pnp-client/requirements.txt
netifaces==0.11.0
python-decouple==3.5 
requests==2.28.1
